requirements.txt
ckanapi
xmltodict
owslib
pyvis
psycopg2-binary
osmium
pyparsing
pyyaml
requests
landez
geopandas
numpy
rasterio
setuptools
gdal
python-dotenv
#pip install --no-deps git+https://github.com/hotosm/osm-export-tool-python.git